requirements.txt
google-cloud-storage
google-cloud-dlp
gunicorn
Flask
Pillow